copilot/chatbot-ai-search-ingest/requirements/requirements.txt
azure-core==1.30.0
azure-identity==1.15.0
azure-search-documents==11.4.0
llama-index==0.10.58
llama-index-agent-openai-legacy==0.1.6
llama-index-core==0.10.58
llama-index-embeddings-azure-openai==0.1.11
llama-index-llms-azure-openai==0.1.10
llama-index-llms-openai==0.1.27
llama-index-postprocessor-rankgpt-rerank==0.1.3
llama-index-storage-chat-store-redis==0.1.3
llama-index-vector-stores-azureaisearch==0.1.13
msal==1.27.0
azure-keyvault-secrets==4.8.0
pandas==2.2.1
pyodbc==5.1.0
requests==2.31.0
python-dotenv==1.0.1
PyJWT==2.10.1
opencensus-ext-azure==1.1.13
sql-metadata==2.11.0
openai==1.24.0
openpyxl==3.1.2
httpx==0.27.2
fastapi==0.115.8
uvicorn==0.32.0